requirements/api-requirements.txt
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter